digitalbank-automation/requirements.txt
# Framework de test
pytest==8.0.0
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2

# Automatisation Mobile/Web
appium-python-client==3.1.0
selenium==4.16.0

# Tests API
requests==2.31.0
jsonschema==4.21.0

# Gestion des données
Faker==22.0.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
PyYAML==6.0.1

# Accessibilité
axe-selenium-python==2.1.6

# Utilitaires
python-dateutil==2.8.2
cryptography==41.0.7

# Reporting
tabulate==0.9.0
